Rework Player actions
